+++ src/toisto/model/language/label.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import re
 from collections.abc import Iterable, Iterator
+from difflib import SequenceMatcher
 from functools import cached_property
 from itertools import chain
 from random import shuffle
@@ -221,6 +222,13 @@ class Label:
         capitonym_key = (self.language, str(self).lower())
         return Labels(label for label in self.capitonym_mapping[capitonym_key] if not self.is_homograph(label))
 
+    def similarity(self, other: Label) -> float:
+        """Return the similarity between this label and the other label as float in the range [0, 1].
+
+        A similarity of 1 means the labels are equal and 0 means they are completely different.
+        """
+        return SequenceMatcher(a=str(self).lower(), b=str(other).lower()).ratio()
+
 
 class Labels:  # noqa: PLW1641
     """Labels collection."""
@@ -269,6 +277,12 @@ class Labels:  # noqa: PLW1641
     def with_same_grammatical_base(self, other: Label) -> Labels:
         """Return the labels with the specified grammatical categories."""
         return Labels(label for label in self._labels if label.has_same_grammatical_base(other))
+
+    def most_similar_label(self, other: Label, min_similarity: float = 0.6) -> Label | None:
+        """Return the most similar label to the other label that has at least the minimum simularity."""
+        if similar_labels := [label for label in self._labels if label.similarity(other) >= min_similarity]:
+            return sorted(similar_labels, key=lambda label: label.similarity(other))[-1]
+        return None
 
     def matching(self, other: Label, *, case_sensitive: bool = True) -> Labels:
         """Return the labels that match the other label."""